Add slug_id field with a RegexValidator
This gets automatically verified in the REST API, **and** in the admin
forms and also is documented in both. Not the use-case I was looking for.

I'm currently struggling hard to find a legitimate use-case that
actually requires the use of the openapi validation I have in mind.

Might be something that actually requires validating two (or more)
fields together, i.e. would normally require python code but can also
be expressed in openapi?

But can't that be expressed in a validator that would be automatically
documented in this way as well?

Also there may be examples that can **not** be expressed in openapi,
but can be expressed in python. In that case the fallback would be to
explain it in the help_text/description and openapi validation would
not work at all.