Add JSONSchemaValidator (prototype)
This

        1. Lists the schema in the migration, very nice side-effect if
        you *know* that Django Migrations are about models, not *only*
        about the database schema.

        2. Validates in the admin form

        3. Validates in DRF post/put. Both validations return the
        appropriate errors.

Missing:

        1. Output of Schema in drf-spectacular.
        2. Output of validation hints in DRF (optional)

2. is highly optional because DRF itself gives very sparse hint in its
on UI about the data types or validation. I'm only going to do it if
a) it's easy and b) if there's an easy way to add that info for other
field types (e.g. List Field) as well.

help_text is probably a better way to achieve this.